+++ src/Components/MainMenuStack/StartStack/MainGameStack/screens/WorkshopScreen.jsx
@@ -1,0 +1,9 @@
+'use client';
+
+import WorkShop from '../WorkShop/WorkShop';
+
+const WorkshopScreen = ({ mapData, navigateToMainGame }) => (
+  <WorkShop navigateToMainGame={navigateToMainGame} mapData={mapData} />
+);
+
+export default WorkshopScreen;